Fix compiler warnings in glance-appwidget
- Remove redundant cast
- Remove redundant conversion call

Test: compile with Kotlin 2.3.0

+++ glance/glance-appwidget/src/main/java/androidx/glance/appwidget/remotecompose/components/RcText.kt
@@ -103,9 +103,7 @@ internal class RcText(
                 colorProvider = colorProvider,
                 fontSizePx = fontSizePx,
                 fontStyle = style?.fontStyle?.style ?: defaultFontStyle, // fontStyle
-                fontWeight =
-                    style?.fontWeight?.value?.toFloat()
-                        ?: defaultFontWeight.toFloat(), // fontWeight
+                fontWeight = style?.fontWeight?.value?.toFloat() ?: defaultFontWeight, // fontWeight
                 fontFamily = style?.fontFamily?.family, // fontFamily TODO, don't hardcode
                 maxLines = emittable.maxLines,
             )
